backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
google-auth
httpx
python-multipart
google-generativeai